fix: navbar, browseStyle, adding component
