+++ src/app/globals.css
@@ -19,6 +19,13 @@
   }
 }
 
+@layer utilities {
+  .font-integral {
+    font-family: 'Integral CF', sans-serif;
+  }
+}
+
+
 body {
   background: var(--background);
   color: var(--foreground);
