+++ src/app/layout.js
@@ -1,7 +1,7 @@
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import Navbar from "./components/Navbar"; // ✅ adjust path if needed
-import Footer from "./components/Footer";
+import Navbar from "./components/Navbar"; 
+import Footer from "./components/Foot";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
